pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cs2-analyzer"
version = "0.1.0"
requires-python = ">=3.12"
description = "CS2 Skins Analysis System — decision engine for buy/no-buy/review"

dependencies = [
    "httpx>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "tomli>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "pytest-cov>=5.0",
]

[project.scripts]
cs2 = "cs2.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
